Move conda support files next to the code
Better than in a separate files dir IMO
