+++ repo2docker/buildpacks/conda/__init__.py
@@ -3,7 +3,7 @@ Generates a variety of Dockerfiles based on an input matrix
 """
 from traitlets import default
 import os
-from .base import BuildPack
+from ..base import BuildPack
 
 
 class CondaBuildPack(BuildPack):
